Method Class Assignment Submission
Create a class. In that class, create a void method that takes two integers as parameters. Have the method do a math operation on the first integer and display the second integer to the screen.

In the Main() method of the console app, instantiate the class.

Call the method in the class, passing in two numbers.

Call the method in the class, specifying the parameters by name.