Lambda Expression Assignment
Create an Employee class with the following properties:

a. Id

b. First Name

c. Last Name

In the Main() method, create a list of at least 10 employees. At least two employees should have the first name “Joe”.

Using a foreach loop, create a new list of all employees with the first name “Joe”. In your comparison statement, remember to reference the property of the object you are checking.

Perform the same action again, but this time with a lambda expression.

Using a lambda expression, make a list of all employees with an Id number greater than 5